Changed up some formatting.

+++ source/Boondocks.Cli/TemporaryFile.cs
@@ -1,19 +1,19 @@
-﻿using System;
-using System.IO;
-
-namespace Boondocks.Cli
+﻿namespace Boondocks.Cli
 {
+    using System;
+    using System.IO;
+
     /// <summary>
-    /// Doesn't actually create the temporary file, but makes sure that it gets deleted after use.
+    ///     Doesn't actually create the temporary file, but makes sure that it gets deleted after use.
     /// </summary>
     internal class TemporaryFile : IDisposable
     {
+        private bool _isDisposed;
+
         public TemporaryFile(string path = null)
         {
             Path = path ?? System.IO.Path.Combine(System.IO.Path.GetTempPath(), $"{Guid.NewGuid():D}.tmp");
         }
-
-        private bool _isDisposed;
 
         public string Path { get; }
 
@@ -24,10 +24,7 @@ namespace Boondocks.Cli
 
             _isDisposed = true;
 
-            if (File.Exists(Path))
-            {
-                File.Delete(Path);
-            }
+            if (File.Exists(Path)) File.Delete(Path);
         }
     }
 }
